src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["jrdalarmi <jeferson.dalarmi@shipay.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
